Add input fetching to 'init' helper binary
Now, the `init` workspace crate can use your session token cookie to
fetch your input from AoC. It checks if the file is already downloaded,
skipping if so. If you do not specify a particular day to download, it
will attempt to download (still checking for already-downloaded inputs)
the inputs for all present day runners for the specified year.